Revert "♻️ refactor: folder struct and remove useless css & map file"
