+++ src/components/Company/components/DashBoard.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './css/DashBoard.scss';
+import '../css/DashBoard.css';
 import Financial from './Financial';
 import RecommendTrend from './RecommendTrend';
 import Earning from './Earning';
