+++ src/components/Company/components/Earning.jsx
