+++ src/components/Company/css/Company.css
@@ -1,0 +1,91 @@
+.container {
+  height: calc(100vh - 96px);
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  padding: 32px;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  font-size: 1vw;
+}
+
+@media screen and (max-width: 820px) {
+  .container {
+    width: 100%;
+    height: 100%;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+        -ms-flex-direction: column;
+            flex-direction: column;
+    padding: 24px 24px;
+  }
+}
+
+.container .company-chart-view {
+  width: 70%;
+  height: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+}
+
+@media screen and (max-width: 820px) {
+  .container .company-chart-view {
+    width: 100%;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: reverse;
+        -ms-flex-direction: column-reverse;
+            flex-direction: column-reverse;
+  }
+}
+
+.container .company-chart-view .company-chart {
+  width: 65%;
+  height: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  margin-left: 10px;
+}
+
+.container .company-chart-view .company-chart .company-title {
+  background-color: white;
+  margin-bottom: 10px;
+}
+
+@media screen and (max-width: 820px) {
+  .container .company-chart-view .company-chart .company-title {
+    font-size: 2em;
+  }
+}
+
+@media screen and (max-width: 820px) {
+  .container .company-chart-view .company-chart {
+    width: 100%;
+    height: 400px;
+    margin-left: 0px;
+  }
+}
+
+.container .vertical-line {
+  width: 1px;
+  background-color: #0c2a6b;
+  margin: 20px;
+  border-radius: 2px;
+}
+
+@media screen and (max-width: 820px) {
+  .container .vertical-line {
+    display: none;
+  }
+}
+/*# sourceMappingURL=Company.css.map */
