+++ src/components/Company/css/Company.scss
