+++ src/components/Company/css/News.css
@@ -1,0 +1,83 @@
+.dashboard-news-list {
+  padding: 10px;
+}
+
+.dashboard-news-list a {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  border-bottom: 1px solid #cecece;
+}
+
+.dashboard-news-list a:first-child {
+  border-top: 1px solid #cecece;
+}
+
+.dashboard-news-list a:hover {
+  background-color: #e3f4fd;
+}
+
+.dashboard-news-list a .dashboard-news-item {
+  width: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  padding: 10px;
+}
+
+.dashboard-news-list a .dashboard-news-item div {
+  width: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  text-align: left;
+  overflow: hidden;
+}
+
+.dashboard-news-list a .dashboard-news-item div .dashboard-news-title {
+  font-size: 1.1em;
+  font-weight: bold;
+  padding-right: 20px;
+}
+
+.dashboard-news-list a .dashboard-news-item div .dashboard-news-date {
+  text-align: left;
+  font-size: 0.7em;
+}
+
+.dashboard-news-list a .dashboard-news-item img {
+  width: 25%;
+}
+
+.dashboard-news-list .dashboard-news-more-btn {
+  width: 100%;
+  height: 40px;
+  margin-top: 15px;
+}
+
+.dashboard-news-list .dashboard-news-more-btn button {
+  width: 100%;
+  height: 100%;
+  border-radius: 5px;
+  background-color: #0084ff;
+  color: white;
+  font-weight: bold;
+}
+
+.dashboard-news-list .dashboard-news-more-btn button:hover {
+  background-color: #0070da;
+}
+/*# sourceMappingURL=News.css.map */
