+++ src/components/Company/css/News.scss
