+++ src/components/Company/css/Profile.scss
