+++ src/components/Header/css/index.css
@@ -1,0 +1,222 @@
+#root .haeder-wrap {
+  width: 100%;
+}
+
+#root .haeder-wrap .header-top {
+  height: 64px;
+  background-color: #0084ff;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  padding: 0 40px;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-top {
+    height: 48px;
+    padding: 0 20px;
+  }
+}
+
+#root .haeder-wrap .header-top .logo {
+  width: 150px;
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-top .logo {
+    width: 100px;
+  }
+}
+
+#root .haeder-wrap .header-top .input-box {
+  position: relative;
+}
+
+#root .haeder-wrap .header-top .input-box input.search-input {
+  width: 30vw;
+  padding: 12px 60px 12px 20px;
+  border: none;
+  border-radius: 8px;
+}
+
+#root .haeder-wrap .header-top .input-box input.search-input:focus {
+  outline: none;
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-top .input-box input.search-input {
+    display: none;
+  }
+}
+
+#root .haeder-wrap .header-top .input-box .search-icon {
+  width: 24px;
+  height: 24px;
+  position: absolute;
+  right: 20px;
+  top: calc(50% - 12px);
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-top .input-box .search-icon {
+    width: 16px;
+    height: 16px;
+    right: 10px;
+    top: calc(50% - 8px);
+  }
+}
+
+#root .haeder-wrap .header-bottom {
+  width: 100%;
+  height: 32px;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  background-color: #02005f;
+  -webkit-box-shadow: 2px 2px 4px 0 rgba(0, 0, 0, 0.15);
+          box-shadow: 2px 2px 4px 0 rgba(0, 0, 0, 0.15);
+  overflow: hidden;
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-bottom {
+    height: 24px;
+  }
+}
+
+#root .haeder-wrap .header-bottom .banner {
+  top: 0;
+  height: 32px;
+  line-height: 32px;
+  background-color: inherit;
+  font-size: 16px;
+  white-space: nowrap;
+  color: white;
+  -webkit-animation: moveLeft 90s linear infinite;
+          animation: moveLeft 90s linear infinite;
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-bottom .banner {
+    height: 24px;
+    background-color: inherit;
+    line-height: 24px;
+    font-size: 12px;
+  }
+}
+
+#root .haeder-wrap .header-bottom .banner .banner-ticker {
+  margin-left: 10px;
+}
+
+#root .haeder-wrap .header-bottom .banner .red-text {
+  margin-left: 5px;
+  color: #ff4141;
+}
+
+#root .haeder-wrap .header-bottom .banner .blue-text {
+  margin-left: 5px;
+  color: #0084ff;
+}
+
+#root .haeder-wrap .header-bottom .banner2 {
+  top: 0;
+  height: 32px;
+  line-height: 32px;
+  background-color: inherit;
+  color: white;
+  font-size: 16px;
+  white-space: nowrap;
+  -webkit-animation: moveLeft2 90s linear infinite;
+          animation: moveLeft2 90s linear infinite;
+}
+
+@media screen and (max-width: 820px) {
+  #root .haeder-wrap .header-bottom .banner2 {
+    height: 24px;
+    background-color: inherit;
+    line-height: 24px;
+    font-size: 12px;
+  }
+}
+
+#root .haeder-wrap .header-bottom .banner2 .banner-ticker {
+  margin-left: 10px;
+}
+
+#root .haeder-wrap .header-bottom .banner2 .red-text {
+  margin-left: 5px;
+  color: #ff4141;
+}
+
+#root .haeder-wrap .header-bottom .banner2 .blue-text {
+  margin-left: 5px;
+  color: #0084ff;
+}
+
+@-webkit-keyframes moveLeft {
+  0% {
+    -webkit-transform: translateX(0);
+            transform: translateX(0);
+  }
+  50% {
+    -webkit-transform: translateX(-100%);
+            transform: translateX(-100%);
+  }
+  50.01% {
+    -webkit-transform: translateX(100%);
+            transform: translateX(100%);
+  }
+  100% {
+    -webkit-transform: translateX(0);
+            transform: translateX(0);
+  }
+}
+
+@keyframes moveLeft {
+  0% {
+    -webkit-transform: translateX(0);
+            transform: translateX(0);
+  }
+  50% {
+    -webkit-transform: translateX(-100%);
+            transform: translateX(-100%);
+  }
+  50.01% {
+    -webkit-transform: translateX(100%);
+            transform: translateX(100%);
+  }
+  100% {
+    -webkit-transform: translateX(0);
+            transform: translateX(0);
+  }
+}
+
+@-webkit-keyframes moveLeft2 {
+  0% {
+    -webkit-transform: translateX(0);
+            transform: translateX(0);
+  }
+  100% {
+    -webkit-transform: translateX(-200%);
+            transform: translateX(-200%);
+  }
+}
+
+@keyframes moveLeft2 {
+  0% {
+    -webkit-transform: translateX(0);
+            transform: translateX(0);
+  }
+  100% {
+    -webkit-transform: translateX(-200%);
+            transform: translateX(-200%);
+  }
+}
+/*# sourceMappingURL=index.css.map */
