+++ src/components/Home/IndexList.jsx
@@ -1,4 +1,4 @@
-import StockTileList from 'pages/Stock/StockTileList';
+import StockTileList from 'components/Stock/StockTileList';
 import React, { useState } from 'react';
 
 const IndexList = ({ stockList, setViewStock }) => {
