+++ src/components/Home/style/index.css
@@ -1,0 +1,231 @@
+.home {
+  height: calc(100vh - 96px);
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  padding: 32px;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-flex: 1;
+      -ms-flex-positive: 1;
+          flex-grow: 1;
+  font-size: 1vw;
+}
+
+@media screen and (max-width: 820px) {
+  .home {
+    width: 100%;
+    height: 100%;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+        -ms-flex-direction: column;
+            flex-direction: column;
+    padding: 24px 24px;
+  }
+}
+
+.home .home-chart-view {
+  width: 70%;
+  height: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+}
+
+@media screen and (max-width: 820px) {
+  .home .home-chart-view {
+    width: 100%;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: reverse;
+        -ms-flex-direction: column-reverse;
+            flex-direction: column-reverse;
+  }
+}
+
+.home .home-chart-view .main-chart {
+  width: 65%;
+  height: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  margin-left: 10px;
+}
+
+@media screen and (max-width: 820px) {
+  .home .home-chart-view .main-chart {
+    margin-left: 0px;
+  }
+}
+
+.home .home-chart-view .main-chart .main-title {
+  background-color: white;
+  margin-bottom: 10px;
+}
+
+@media screen and (max-width: 820px) {
+  .home .home-chart-view .main-chart .main-title {
+    font-size: 2em;
+  }
+}
+
+@media screen and (max-width: 820px) {
+  .home .home-chart-view .main-chart {
+    width: 100%;
+    height: 400px;
+  }
+}
+
+.home .line {
+  width: 1px;
+  background-color: #0c2a6b;
+  margin: 20px;
+  border-radius: 2px;
+}
+
+@media screen and (max-width: 820px) {
+  .home .line {
+    display: none;
+  }
+}
+
+.home .right-area {
+  width: 30%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  background-color: white;
+  padding: 20px;
+}
+
+@media screen and (max-width: 820px) {
+  .home .right-area {
+    margin-top: 10px;
+    width: 100%;
+  }
+}
+
+.home .right-area .side-title button {
+  margin-right: 10px;
+  font-size: 1.2em;
+  padding: 5px 10px;
+  border-radius: 5px;
+  font-weight: bold;
+}
+
+@media screen and (max-width: 820px) {
+  .home .right-area .side-title button {
+    font-size: 3em;
+  }
+}
+
+.home .right-area .side-title .unselected:hover {
+  background-color: #e9e9e9;
+}
+
+.home .right-area .side-title .selected {
+  background-color: #0084ff;
+  color: white;
+}
+
+.home .right-area .stock-list {
+  height: 100%;
+  overflow-y: hidden;
+  font-size: 0.65em;
+  margin-top: 10px;
+  padding-right: 10x;
+}
+
+@media screen and (max-width: 820px) {
+  .home .right-area .stock-list {
+    width: 100%;
+    overflow-y: auto;
+  }
+}
+
+.home .right-area .stock-list a {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  height: calc(100% / 5);
+  border-bottom: 1px solid #cecece;
+}
+
+.home .right-area .stock-list a:first-child {
+  border-top: 1px solid #cecece;
+}
+
+.home .right-area .stock-list a:hover {
+  background-color: #e3f4fd;
+}
+
+.home .right-area .stock-list .trending-item {
+  width: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  text-align: left;
+}
+
+@media screen and (max-width: 820px) {
+  .home .right-area .stock-list .trending-item {
+    font-size: 2.2em;
+  }
+}
+
+.home .right-area .stock-list .news-item {
+  width: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  padding: 10px;
+}
+
+.home .right-area .stock-list .news-item div {
+  width: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  text-align: left;
+  overflow: hidden;
+}
+
+.home .right-area .stock-list .news-item div .news-title {
+  font-size: 1.7em;
+  font-weight: bold;
+  padding-right: 20px;
+}
+
+.home .right-area .stock-list .news-item div .news-date {
+  text-align: left;
+}
+
+.home .right-area .stock-list .news-item img {
+  width: 25%;
+}
+/*# sourceMappingURL=index.css.map */
