+++ src/components/SearchList/styles/SearchListStyle.scss
