+++ src/components/Stock/StockDetailChart.jsx
