+++ src/components/Stock/StockItem.jsx
