+++ src/components/Stock/StockTileList.jsx
