+++ src/components/Stock/css/StockDetailChart.css
@@ -1,0 +1,114 @@
+.chart-container {
+  height: 100%;
+  background-color: white;
+}
+
+.chart-container .period-selector {
+  height: 50px;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: justify;
+      -ms-flex-pack: justify;
+          justify-content: space-between;
+}
+
+.chart-container .period-selector ul {
+  height: 50px;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+}
+
+.chart-container .period-selector ul li {
+  float: left;
+}
+
+.chart-container .period-selector ul li input {
+  display: none;
+}
+
+.chart-container .period-selector ul li label {
+  background-color: white;
+  margin: 5px 5px;
+  margin-left: 10px;
+  padding: 5px 10px;
+  cursor: pointer;
+  border-radius: 5px;
+  font-weight: bold;
+}
+
+.chart-container .period-selector ul li label:hover {
+  background-color: #e9e9e9;
+}
+
+@media screen and (max-width: 820px) {
+  .chart-container .period-selector ul li label {
+    font-size: 15px;
+  }
+}
+
+.chart-container .period-selector ul li input:checked + label {
+  background-color: #0084ff;
+  color: white;
+  font-weight: bold;
+}
+
+.chart-container .period-selector ul li button {
+  width: 30px;
+  height: 30px;
+  margin-left: 10px;
+  border-radius: 5px;
+}
+
+.chart-container .period-selector ul li button:hover {
+  background-color: #e9e9e9;
+}
+
+.chart-container .period-selector ul li button svg {
+  width: 20px;
+  height: 20px;
+  vertical-align: middle;
+  fill: #0084ff;
+}
+
+.chart-container .period-selector .update-time {
+  height: 100%;
+  line-height: 50px;
+  padding-right: 20px;
+}
+
+.chart-container .period-selector .update-time img {
+  width: 17px;
+  height: 17px;
+  vertical-align: middle;
+  background-color: #dbdbdb;
+  border-radius: 17px;
+  padding: 2px;
+  border: 1px solid #dbdbdb;
+}
+
+.chart-container .period-selector .update-time span {
+  color: #999999;
+  vertical-align: middle;
+  font-size: 12px;
+}
+
+.chart-container .chart {
+  width: 100%;
+  height: calc(100% - 50px);
+  padding: 10px;
+}
+
+.chart-container .chart .tooltip-box {
+  background-color: white;
+  padding: 10px;
+}
+
+.chart-container .chart .tooltip-box span {
+  font-weight: bold;
+}
+/*# sourceMappingURL=StockDetailChart.css.map */
