+++ src/components/Stock/css/StockDetailChart.scss
