+++ src/components/Stock/css/StockItem.scss
