+++ src/components/Stock/css/StockTileList.css
@@ -1,0 +1,51 @@
+.tile-list {
+  width: 35%;
+  height: 100%;
+  font-size: 0.7vw;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+}
+
+@media screen and (max-width: 820px) {
+  .tile-list {
+    width: 100%;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+        -ms-flex-direction: row;
+            flex-direction: row;
+    margin-top: 10px;
+    padding-bottom: 10px;
+    overflow-x: auto;
+    font-size: 1.1vw;
+  }
+}
+
+.tile-list button {
+  height: 100%;
+  background-color: white;
+  text-align: left;
+  margin-top: 10px;
+}
+
+.tile-list button:first-child {
+  margin-top: 0px;
+}
+
+.tile-list button:hover {
+  background-color: #e3f4fd;
+}
+
+@media screen and (max-width: 820px) {
+  .tile-list button {
+    height: auto;
+    min-width: 200px;
+    margin-right: 5px;
+    margin-top: 0px;
+  }
+}
+/*# sourceMappingURL=StockTileList.css.map */
