+++ src/components/Stock/css/StockTileList.scss
